requirements.txt
numpy~=1.21.0
scipy~=1.7.2
mne
matplotlib~=3.5.0
renalabapp
pyxdf~=1.16.3'
pandas
torchsummary==1.5.1
